Cityscapes related improvements

+++ data/cityscapes_generator.py
@@ -178,9 +178,13 @@ class CityscapesGenerator(Sequence):
 
             # Initialize image_batch and label_batch if needed
             if image_batch is None:
-                image_batch = np.empty((self.batch_size, ) + image.shape)
+                image_batch = np.empty(
+                    (self.batch_size, ) + image.shape, dtype=np.uint8
+                )
             if label_batch is None:
-                label_batch = np.empty((self.batch_size, ) + label.shape)
+                label_batch = np.empty(
+                    (self.batch_size, ) + label.shape, dtype=np.uint8
+                )
 
             # Fill image_batch and label_batch iteratively
             image_batch[idx] = image
